@@ -29,82 +30,54 @@
         "packages/perspective-jupyterlab",
         "packages/perspective-webpack-plugin",
         "packages/perspective-cli",
-        "examples/blocks",
+        "examples/*",
         "docs",
         "python/perspective"
     ],
     "devDependencies": {
         "@types/ws": "^7.2.2",
-        "arraybuffer-loader": "^1.0.2",
+        "auto-changelog": "^2.4.0",
         "chalk": "^2.4.2",
-        "cpy-cli": "^3.1.1",
-        "css-loader": "^0.28.7",
-        "cssnano": "^4.1.10",
-        "cssnano-preset-lite": "^1.0.1",
+        "cpy-cli": "^4.2.0",
+        "css-loader": "^6.7.3",
         "decompress": "^4.2.1",
         "decompress-unzip": "^4.0.1",
         "dotenv": "^8.1.0",
-        "esbuild": "^0.13.12",
-        "esbuild-plugin-less": "^1.1.5",
+        "esbuild": "^0.14.54",
+        "esbuild-plugin-less": "^1.1.11",
         "fs-extra": "^8.1.0",
-        "html-loader": "^0.5.1",
+        "glob": "^7.2.0",
         "html-loader-jest": "^0.2.1",
         "html-webpack-plugin": "^5.1.0",
         "husky": "^7.0.4",
         "inquirer": "^7.0.0",
-        "jest": "^29.3.1",
+        "jest": "^25.1.0",
         "jest-junit": "^10.0.0",
-        "jsdoc": "3.5.5",
-        "jsdoc-babel": "^0.5.0",
-        "jsdoc-to-markdown": "^4.0.1",
-        "less": "^4.1.0",
+        "jsdoc-to-markdown": "^8.0.0",
+        "less": "^4.1.3",
         "less-loader": "^7.2.1",
-        "mini-css-extract-plugin": "^1.3.4",
+        "minimatch": "^5.1.2",
         "minimist": "^1.2.0",
         "mkdirp": "^0.5.1",
         "npm-pack-here": "^1.2.0",
         "npm-run-all": "^4.1.3",
         "octokit": "^1.7.2",
-        "postcss": "^8",
-        "postcss-loader": "^7",
         "prettier": "^2.7.1",
         "puppeteer": "^13.1.3",
         "rimraf": "^2.6.2",
         "sinon": "^7.3.1",
-        "style-loader": "^0.18.2",
+        "style-loader": "^3.3.1",
         "term-img": "^4.1.0",
+        "timezone-mock": "^1.3.6",
         "tiny-glob": "^0.2.9",
         "ts-loader": "^6.2.0",
-        "typedoc": "^0.22.7",
-        "typedoc-plugin-markdown": "^3.11.6",
-        "typedoc-plugin-no-inherit": "^1.3.1",
-        "typescript": "4.3.3",
-        "webfontloader": "^1.6.28",
+        "typedoc": "^0.23.0",
+        "typedoc-plugin-markdown": "^3.14.0",
+        "typedoc-plugin-no-inherit": "^1.4.0",
+        "typescript": "4.9.4",
         "webpack": "^5.60.0",
         "webpack-cli": "^4.3.1",
-        "webpack-dev-server": "^3.11.2",
-        "worker-loader": "^3.0.7"
-    },
-    "resolutions": {
-        "graceful-fs": "^4.2.4",
-        "jest": "^25.1.0",
-        "write-file-atomic": "3.0.0",
-        "postcss": "^8",
-        "webpack": "^5.60.0",
-        "react": "16.8.6",
-        "react-dom": "16.8.6",
-        "npm-packlist": "1.1.12",
-        "@lumino/algorithm": "^1.9.1",
-        "@lumino/commands": "^1.20.0",
-        "@lumino/coreutils": "^1.12.0",
-        "@lumino/disposable": "^1.10.1",
-        "@lumino/domutils": "^1.8.1",
-        "@lumino/dragdrop": "^1.14.0",
-        "@lumino/keyboard": "^1.8.1",
-        "@lumino/messaging": "^1.10.1",
-        "@lumino/properties": "^1.8.1",
-        "@lumino/signaling": "^1.10.1",
-        "@lumino/virtualdom": "^1.14.1"
+        "webpack-dev-server": "^4.11.1"
     },
     "scripts": {
         "postinstall": "node scripts/install_emsdk.js",
