@@ -18,6 +18,8 @@ const {
     NodeModulesExternal,
 } = require("@finos/perspective-esbuild-plugin/external");
 
+const cpy_mod = import("cpy");
+
 const IS_DEBUG =
     !!process.env.PSP_DEBUG || process.argv.indexOf("--debug") >= 0;
 
@@ -104,8 +106,10 @@ async function build_all() {
     await Promise.all(POSTBUILD.map(build)).catch(() => process.exit(1));
 
     // legacy compat
-    execSync("cpy target/themes/* dist/css");
-    execSync("cpy dist/css/* dist/umd");
+    const { default: cpy } = await cpy_mod;
+    cpy("target/themes/*", "dist/css");
+    cpy("target/themes/*", "dist/umd");
+    cpy("dist/pkg/*", "dist/esm");
 }
 
 build_all();
