@@ -10,6 +10,7 @@
 
 import type { IPerspectiveViewerElement } from "./viewer";
 import type { HTMLPerspectiveViewerPluginElement } from "./plugin";
+import type React from "react";
 
 // JSX / React extensions
 
@@ -50,7 +51,7 @@ declare global {
     }
 }
 
-import { PerspectiveViewerElement } from "@finos/perspective-viewer/dist/pkg/perspective.js";
+import { PerspectiveViewerElement } from "../../dist/pkg/perspective.js";
 
 // This unused class asserts that the _real_ class (the one generated by
 // `wasm-bindgen`) implements the same interface we register to the
